+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# V 1.1.0
+Changes in this release:
+- Added --use-module-in-place option (#30)
+- Added support to test regular functions and classes (#37)
+- Close handler caches on cleanup (#42)
+
 # V 1.0.0
 Changes in this release:
 - Added support to get logs from handler (#35)